==> code/01 Data_Collection.Rmd <==
---
title: "01 Data Collection"
author: "Daniel Tshiani"
date: "2025-10-07"
output: html_document
---

```{r}
library(itscalledsoccer)
library(dplyr)
library(keyring)
library(httr2)
library(tidyverse)
library(jsonlite)
library(purrr)
library(lubridate)
```

```{r}
# retrieving the environment
data <- AmericanSoccerAnalysis$new()
```

```{r}
# retrieving data
players_data <- data$players
team_data <- data$teams
player_salary_data <- data $get_player_salaries()
player_goals_added_data <- data$get_player_goals_added()
player_xpass_data <- data$get_player_xpass()
player_xgoals_data <- data$get_player_xgoals()
rm(data)
```

```{r}
# Creating master player-level dataset by season

master_data <- player_salary_data
glimpse(master_data)
```

```{r}
sum(duplicated(master_data[, c("player_id", "team_id", "season_name", "base_salary", "guaranteed_compensation", "competition")]))

# there are duplicates based on mlspa_release
sum(duplicated(master_data[, c("player_id", "team_id", "season_name", "mlspa_release")]))

# keeping the later salaries by mlspa_release
master_data <- master_data %>%
  arrange(player_id, team_id, season_name, mlspa_release)

master_data <- master_data %>%
  distinct(player_id, team_id, season_name, .keep_all = TRUE) %>%
  select(-mlspa_release)

rm(player_salary_data)
glimpse(master_data)
```

```{r}
# joining player data
glimpse(players_data)
sum(duplicated(players_data$player_id))
```
```{r}
unmatched_in_master <- anti_join(master_data, players_data, by = "player_id") # everyone is the master data appears in player data
unmatched_in_player <- anti_join(players_data, master_data, by = "player_id") # there are players in the player data who arent in the master. regardless in the end i want all the players so i'll keep them and only dropped them during model training.
rm(unmatched_in_master, unmatched_in_player)
master_data <- full_join(master_data, players_data, by= "player_id")
rm(players_data)
```

```{r}
# joining team data
glimpse(team_data)
team_data <- team_data %>%
  mutate(team_competitions = competitions) %>%
  select(-competitions)

master_data <- full_join(master_data, team_data, by = "team_id")
rm(team_data)
```

```{r}
# Given the timeframe, I am excluding the player goals added data. I didnt understand the meaning of the values. in future iterations on the app this data should be looked into again.
glimpse(player_goals_added_data)
rm(player_goals_added_data)
```

```{r}
glimpse(player_xgoals_data)
sum(duplicated(player_xgoals_data$player_id))
sum(duplicated(master_data$player_id))

# need to import it 1 season at a time and label the season_name in order to join it properly with master data
data <- AmericanSoccerAnalysis$new()

for (year in 2013:2025) {
  var_name <- paste0("player_xpass_data_", year)
  
  temp <- data$get_player_xpass(season_name = as.character(year)) %>%
    mutate(season_name = year)
  
  assign(var_name, temp)
}
#now i can run a loop to do a full join by player_id and season_name; and then clean up the environment
player_xpass_all <- bind_rows(
  lapply(2013:2025, function(year) get(paste0("player_xpass_data_", year)))
)

rm(list = paste0("player_xpass_data_", 2013:2025), player_xpass_data)
test <- full_join(master_data, player_xpass_all, by= c("player_id", "season_name", "competition"))

test <- test%>%
  mutate(match = ifelse(team_id.x == team_id.y, 1, 0))
table(test$match) # I viewed the data and team_id.y is a list so i will prioritize team_id.x first
rm(test)
master_data <- full_join(master_data, player_xpass_all, by= c("player_id", "season_name", "competition"))
rm(player_xpass_all, temp)
```

```{r}
glimpse(player_xgoals_data)
```
```{r}
for (year in 2013:2025) {
  var_name <- paste0("player_xgoals_data_", year)
  
  temp <- data$get_player_xgoals(season_name = as.character(year)) %>%
    mutate(season_name = year)
  
  assign(var_name, temp)
}
#now i can run a loop to do a full join by player_id and season_name; and then clean up the environment
player_xgoals_all <- bind_rows(
  lapply(2013:2025, function(year) get(paste0("player_xgoals_data_", year)))
)

rm(list = paste0("player_xgoals_data_", 2013:2025), player_xgoals_data)
```

```{r}
master_data <- full_join(master_data, player_xgoals_all, by= c("player_id", "season_name", "competition"))
rm(test)
rm(player_xgoals_all, temp, data)
```

# Initial Preprocessing Completed
💡 Summary workflow after merging your dataframe:
Inspect & handle missing values
Remove duplicates
Correct data types
Identify & handle outliers
Feature engineering
Encode categorical variables
Scale if necessary
Split train/test
Explore correlations & distributions
Save cleaned dataset + preprocessing pipeline



# cleaning
```{r}
# i need to do this once everything is merged
# quick clean of player data
test <- players_data %>%
  mutate(
    birth_date = ymd(birth_date),
    age = year(today()) - year(birth_date) - (yday(today()) < yday(birth_date))
  ) %>%
  select(-birth_date) %>%
  relocate(age, .after = player_name)

master_data <- master_data %>%
  mutate(height_inches = height_ft * 12 + height_in) %>%
  select(-height_ft, -height_in) %>%
  relocate(height_inches, .after = age)

master_data <- master_data %>%
  mutate(homegrown = ifelse(nationality %in% c("USA", "Canada"), "Domestic", "Foreign")) %>%
  mutate(homegrown = as.factor(homegrown)) %>%
  relocate (homegrown, .after = nationality) %>%
  relocate(weight_lb, .after = height_inches)
  
glimpse(master_data)
rm(players_data)
```

```{r}
# joining player expected goals data
glimpse(player_xgoals_data)
sum(is.na(player_xgoals_data$player_id))
```
```{r}
unmatched <- bind_rows(
  anti_join(master_data, player_xgoals_data, by = "player_id"),
  anti_join(player_xgoals_data, master_data, by = "player_id")
)
```


```{r}
master_data1 <- left_join(master_data, player_xgoals_data, by = "player_id")
glimpse(master_data1)
```




# 0ld
```{r}
players <- data$get_players()
players2 <- data$players

teams <- data$get_teams()
teams2 <- data$teams

leagues <- data$LEAGUES

player_salary <- data$get_player_salaries()
team_salary <- data$get_team_salaries()


```

```{r}
player_data <- left_join(player_salary, players2, by = "player_id") # joining player and salary data at the season/salary level

player_data <- left_join(player_data, teams2, by = "team_id") # adding team names to the data
```

```{r}
team_salary <- team_salary %>%
  select(-count_players, -std_dev_guaranteed_compensation) %>%
  rename(team_guaranteed_compensation = total_guaranteed_compensation,
         team_avg_gauranteed_compenesation = avg_guaranteed_compensation,
         team_median_guaranteed_compensation = median_guaranteed_compensation,
         team_competition = competition)

player_data <- left_join(player_data, team_salary, by = "team_id") # adding team salary data
```

```{r}
str(data)
```

```{r}
goals <- data$get_player_xgoals(season_name = 2024)
goals_added <- data$get_player_goals_added()
pass <- data$get_player_xpass(split_by_seasons = TRUE)
```

# Focusing on dat for Offensive tab

```{r}
offensive_data <- data$get_player_xgoals(season_name = 2024)
player_data <- data$players
player_data <- player_data %>%
  select(-competitions)

offensive_data <- left_join(offensive_data, player_data, by = "player_id")
offensive_data <- offensive_data %>%
  relocate(player_name, birth_date, general_position, nationality, player_id, team_id)

team_data <- data$teams

offensive_data <- offensive_data %>%
  mutate(team_id = sapply(team_id, function(x) as.character(x[1])))

offensive_data <- left_join(offensive_data, team_data, by = "team_id")
offensive_data <- offensive_data %>% select(-team_short_name, -team_abbreviation)

offensive_data <- offensive_data %>%
  relocate(player_name, team_name, birth_date, general_position, nationality, player_id, team_id)
```


```{r}
offensive_data <- offensive_data %>%
  mutate(
    birth_date = ymd(birth_date),      # convert to Date format if not already
    age = as.integer(floor(interval(birth_date, today()) / years(1)))
  ) %>%
    relocate(player_name, team_name, age, birth_date, general_position, nationality)
```

```{r}
# Adding salary of offensive data
salary_data <- data$get_player_salaries(season_name = 2024)

# getting rid of duplicates player_id and team_id combo
salary_data <- salary_data %>%
  arrange(desc(mlspa_release)) %>%
  distinct(player_id, team_id, .keep_all = TRUE)

offensive_data <- left_join(offensive_data, salary_data, by = c("player_id", "team_id"))
```


```{r}
#saveRDS(offensive_data, file = "../data/offensive_data.rds")
```
