==> code/02 Cleaning.Rmd <==
---
title: "02 Cleaning"
author: "Daniel Tshiani"
date: "2025-10-20"
output: html_document
---

```{r}
library(tidyverse)
```


```{r}
master_data <- readRDS("../data/01 preprocessed data.rds")
glimpse(master_data)
```

# Correcting data types, Encode categorical variables & scaling necessary when
```{r}
master_data <- master_data %>%
  mutate(position = as.factor(position),
         competition = as.factor(competition),
         birth_date = ymd(birth_date),
         age = year(today()) - year(birth_date) - (yday(today()) < yday(birth_date)),
         age = as.integer(age)
         ) %>%
  relocate(age, .after = birth_date) %>%
  mutate(homegrown = ifelse(nationality %in% c("USA", "Canada"), "Domestic", "Foreign")) %>%
  mutate(homegrown = as.factor(homegrown)) %>%
  relocate (homegrown, .after = nationality) %>%
  mutate(height_inches = height_ft * 12 + height_in,
         height_inches = as.integer(height_inches)) %>%
  relocate(height_inches, .after = height_in) %>%
  relocate(competitions, .after = competition) %>% # I need to look into this more closely but competition will be prioritized due to it being a single character vs a list
  relocate(team_competitions, .after = competitions) %>% # need to look into this more closely
  relocate(general_position.x, .after = position) %>%
  relocate(general_position.y, .after = general_position.x) %>%
  mutate(pass_completion_percentage = 100 * pass_completion_percentage,
         xpass_completion_percentage = 100 * xpass_completion_percentage,
         share_team_touches = 100 * share_team_touches)
  
glimpse(master_data)
```

# Checking and Removing duplicates
```{r}
# sum(duplicated(master_data)) # no duplicates across all rows, in an update the competition variables became lists
sum(duplicated(master_data[, c("player_id", "team_id.x", "season_name")])) # 339 duplicates across these 3 variables

duplicates <- master_data %>%
  group_by(player_id, team_id.x, season_name) %>%
  mutate(count = n()) %>%
  filter(n() > 1) %>%   
  ungroup() %>%
  arrange(player_id, team_id.x, season_name)

table(duplicates$team_id.x) # the duplicates are because they are missing the duplicates$team_id.x variable
# looks like one of the previous datasets didnt merge properly. they are all missing team_id.x so i will use team_id.y for these observations

sum(duplicated(duplicates[, c("player_id", "team_id.y", "season_name")])) # when i use the team_id.y variable they are unique
```

```{r}
# to find where the merge is going wrong i need to go back and create flags for each dataset before merging but for the sake of time i will continue to move forward by replacing the missing team_id.x with team_id.y and seeing what i can salvage

class(duplicates$team_id.y) # its a list
max(sapply(duplicates$team_id.y, length)) # max number of elements is 1 so shouldnt cause any big issues

missing_playerID <- unique(duplicates$player_id)

max(sapply(master_data$team_id.y, length))

master_data <- master_data %>%
  mutate(team_id.y2 = map(team_id.y, ~ if (length(.x) == 1) .x else NULL))
max(sapply(master_data$team_id.y2, length)) 

master_data <- master_data %>%
  mutate(team_id.x = ifelse(is.na(team_id.x), team_id.y2, team_id.x)) %>%
  select(-team_id.y2)
```

```{r}
# now that i replaced the missing team_id.x with team_id.y, i will try again to Check and Remove duplicates
#sum(duplicated(master_data[, c("player_id", "team_id.x", "season_name")])) # now i dropped from 339 to 44 duplicates across these 3 variables. again variables were updated to list

duplicates <- master_data %>%
  group_by(player_id, team_id.x, season_name) %>%
  mutate(count = n()) %>%
  filter(n() > 1) %>%   
  ungroup() %>%
  arrange(player_id, team_id.x, season_name)

sum(duplicated(duplicates[, c("player_id", "team_id.y", "season_name")])) 

# I did a visual inspection and looks like these are genuine duplicates. however for each pair, 1 of them is missing team name and related information so i will keep the more complete row in the master_data

master_data <- master_data %>%
  arrange(player_id, season_name, team_id.x, team_name) %>%
  distinct(player_id, season_name, team_id.x, .keep_all = TRUE)
```

# Inspecting & Handling Missing Values
```{r}
library(visdat)
vis_miss(master_data, warn_large_data = FALSE)
```
There is alot of data missing. I dont want to drop players from the final app so i will have 1 dataset for the app will all the data possible, and another cleaner version of the data that will be use to train and test the models

## I want to see how missingsness looks by competition

```{r}
mls <- master_data %>%
  filter(competition == "mls")
vis_miss(mls)
```

```{r}
mlsnp <- master_data %>%
  filter(competition == "mlsnp")
vis_miss(mlsnp)
```
So i don't have any salary data for MLSNP. This is an issue. I am assuming other leagues will be missing salary too but for the purpose of using this project to display my soccer analytics abilities i will continue.

```{r}
nasl <- master_data %>%
  filter(competition == "nasl")
vis_miss(nasl)
```


```{r}
nwsl <- master_data %>%
  filter(competition == "nwsl")
vis_miss(nwsl)
```


```{r}
usl1 <- master_data %>%
  filter(competition == "usl1")
vis_miss(usl1)
```


```{r}
table(master_data$competition)
```
So I am missing alot of data outside the MLS. This is an issue. I will train/test the models with the data i have. I will create another dataset for app deployment so all the players can still be searched.

```{r}
app_deployment_data <- master_data # data for the deployment
working_data <- master_data # data for modeling

rm(duplicates, master_data, mls, mlsnp, nasl, nwsl, usl1, missing_playerID)
```

```{r}
working_data %>%
  summarise(
    pct_missing_base_salary = mean(is.na(base_salary)) * 100
  )
```

```{r}
working_data %>%
  group_by(competition) %>%
  summarise(
    pct_missing_base_salary = mean(is.na(base_salary)) * 100
  )
```

so the plan is basically to drop the variables whos missing base_salary. (I know its not ideal but yea...) so essentially the model will only be train on MLS players

```{r}
working_data <- working_data %>%
  filter(!is.na(base_salary))
```

```{r}
vis_miss(working_data)
```

```{r}
# Now that the missingness througout the dataset is 10%, i will loop through it and replace the missing values with mean or mode depending on the type of variable.

# Function to get the mode
get_mode <- function(x) {
  ux <- na.omit(unique(x))
  ux[which.max(tabulate(match(x, ux)))]
}

working_data <- working_data %>%
  # Drop rows with missing player_id
  filter(!is.na(player_id)) %>%
  
  # Replace missing team_id variants with "Free Agent"
  mutate(across(
    c(team_id.x, team_id.y, team_id),
    ~ replace(.x, is.na(.x) | .x == "", "Free Agent")
  )) %>%
  
  # Replace other missing values
  mutate(across(
    # skip id columns and team columns we already handled
    -c(player_id, team_id.x, team_id.y, team_id),
    ~ if (is.numeric(.x)) {
        replace(.x, is.na(.x), mean(.x, na.rm = TRUE))
      } else {
        replace(.x, is.na(.x), get_mode(.x))
      }
  ))

vis_miss(working_data)
```

# Identify & handle outliers
```{r}
ggplot(working_data, aes(x = base_salary)) +
  geom_histogram(binwidth = 100, fill = "skyblue", color = "black") +
  labs(
    title = "Distribution of Base Salary",
    x = "Base Salary",
    y = "Count"
  ) +
  theme_minimal()
```

```{r}
# Find Messi's salary
messi_salary <- working_data %>%
  filter(player_name == "Lionel Messi", season_name == 2025) %>%
  pull(base_salary)

# Plot
ggplot(working_data, aes(x = base_salary)) +
  geom_histogram(binwidth = 100, fill = "skyblue", color = "black") +
  geom_vline(aes(xintercept = messi_salary),
             color = "red", linewidth = 1.2, linetype = "dashed") +
  annotate("text", x = messi_salary, y = Inf, vjust = 2, hjust = 0,
           label = "Messi", color = "red", fontface = "bold") +
  labs(
    title = "Distribution of Base Salary",
    x = "Base Salary",
    y = "Count"
  ) +
  theme_minimal()

```
based a on a visual inspection there appears to be outliers. However, I decided to leave in the outliers. I knew there are players like Messi who has a crazy salary. I do believe superstar coming to the MLS and having huge salaries is a real phenomenon and should be reflected in the model.

# Cleaning the data structure
I need to clean the structure so ensuring the variables are the right type and name.
```{r}
glimpse(working_data)
```

```{r}
working_data <- working_data %>%
  select(-player_id, -team_id.x) %>%
  rename(season = season_name,
         "general position" = position,
         position = general_position.x,
         "base salary" = base_salary,
         "gauranteed compensation" = guaranteed_compensation) %>%
  select(-general_position.y, -competitions) %>%
  mutate( position = as.factor(position),
          season = as.factor(season)) %>%
  select(-team_competitions) %>%
  rename("player name" = player_name,
         "date of birth" = birth_date,
         "height (ft)" = height_ft,
         "height (in)" = height_in,
         weight = weight_lb) %>%
  select(-height_inches) %>%
  rename("team name" = team_name) %>%
  select(-team_short_name, - team_abbreviation, -team_id.y) %>%
  rename("minutes played" = minutes_played.x,
         "attempted passes" = attempted_passes,
         "pass completion percentage" = pass_completion_percentage,
         "expected pass completion percentage" = xpass_completion_percentage,
         "pass completed over expected" = passes_completed_over_expected,
         "pass completed over expected p100" = passes_completed_over_expected_p100,
         "average distance (yards)" = avg_distance_yds,
         "average distance vertical yards" = avg_vertical_distance_yds,
         "share of team touches" = share_team_touches,
         "games played" = count_games) %>%
  select(-team_id, -minutes_played.y) %>%
  rename("shots on target" = shots_on_target,
         "expected goals" = xgoals,
         "goals - expected goals" = goals_minus_xgoals,
         "key passes" = key_passes,
         "assists" = primary_assists,
         "expected assists" = xassists,
         "primary assists - exp. assists" = primary_assists_minus_xassists,
         "exp. goals + exp. assists" = xgoals_plus_xassists,
         "points added" = points_added,
         "exp.points added" = xpoints_added,
         "goals + assists" = goals_plus_primary_assists)

glimpse(working_data)
```


```{r}
saveRDS(working_data, file = "../data/02 working data.rds") 
```
